update test to use float64 for width and height

+++ image/resize/resizer_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func TestCalculateAspectRatio(t *testing.T) {
 	cases := []struct {
-		width, height int
+		width, height float64
 		aspectRatio   float64
 	}{
 		{100, 100, 1},
@@ -14,15 +14,15 @@ func TestCalculateAspectRatio(t *testing.T) {
 	s := Strategy{}
 	for _, c := range cases {
 		if got := s.calculateAspectRatio(c.width, c.height); got != c.aspectRatio {
-			t.Errorf("calculateAspectRatio(%d, %d) = %f; want %f", c.width, c.height, got, c.aspectRatio)
+			t.Errorf("calculateAspectRatio(%f, %f) = %f; want %f", c.width, c.height, got, c.aspectRatio)
 		}
 	}
 }
 
 func TestCalculateNewDimensions(t *testing.T) {
 	cases := []struct {
-		width, height, maxWidth, maxHeight int
-		expectedWidth, expectedHeight      int
+		width, height, maxWidth, maxHeight float64
+		expectedWidth, expectedHeight      float64
 	}{
 		{100, 100, 200, 200, 200, 200},
 		{200, 100, 0, 0, 200, 100},
@@ -34,7 +34,7 @@ func TestCalculateNewDimensions(t *testing.T) {
 	s := Strategy{}
 	for _, c := range cases {
 		if gotWidth, gotHeight := s.calculateNewDimensions(c.width, c.height, c.maxWidth, c.maxHeight); gotWidth != c.expectedWidth || gotHeight != c.expectedHeight {
-			t.Errorf("calculateNewDimensions(%d, %d, %d, %d) = (%d, %d); want (%d, %d)", c.width, c.height, c.maxWidth, c.maxHeight, gotWidth, gotHeight, c.expectedWidth, c.expectedHeight)
+			t.Errorf("calculateNewDimensions(%f, %f, %f, %f) = (%f, %f); want (%f, %f)", c.width, c.height, c.maxWidth, c.maxHeight, gotWidth, gotHeight, c.expectedWidth, c.expectedHeight)
 		}
 	}
 }
